feat(terraform): add plugin commands for terraform apply and plan
This makes the terraform provider a little easier to work with
day-to-day. We also explicitly discourage use of `autoApply` now, except
for private environments. The status check is now also faster, because
we don't refresh the stack state every time.

This involved a few refactors, mainly to improve the plugin command
framework. I also needed to modify the provider resolution flow so that
individual providers can be resolved (otherwise it wouldn't be possible
to run the apply/plan commands if another provider depended on the
outputs).